Add accessibility note to preflight docs
I'm not sure if I worded this the best way, but the removal of styles from `<ul>` and `<li>` elements has an unexpected side effect, which is that VoiceOver stops treating the content like a list. People using Tailwind need to be mindful of this and I'm hoping a note in the docs will help.